pyproject.toml
[project]
name = "short-story-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "jinja2>=3.1.6",
    "logfire>=4.16.0",
    "pydantic-ai[google]>=1.39.0",
    "python-dotenv>=1.2.1",
    "python-fasthtml>=0.12.36",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
